requirements.txt
numpy==1.22.0
Pillow==9.0.0
Pillow==10.0.0
streamlit==1.21.0
tensorflow==2.11.0
tensorflow_intel==2.11.0
